Phase B4: extract init, reset, serve handlers into src/commands/
