+++ crates/defra-agent-cli/src/commands/mod.rs
@@ -1,1 +1,4 @@
 // Subcommand handler implementations. Populated by Phases B4-B7.
+pub(crate) mod init;
+pub(crate) mod reset;
+pub(crate) mod serve;
